requirements.txt
Flask
flask-restful
celery
urllib3
Jinja2
python-docx
python-memcached
cloudant
